+++ src/http/version/handler_1_0.cpp
@@ -35,7 +35,7 @@ namespace http::version::handler_1_0 {
         if (!conf::matchConfigs.empty()) {
             // Format raw path & remove query string
             std::string rawPath = request.getPathStr();
-            const size_t queryIndex = rawPath.find('/');
+            const size_t queryIndex = rawPath.find('?');
             while (queryIndex != std::string::npos && rawPath.size() > queryIndex)
                 rawPath.pop_back();
 
@@ -64,6 +64,30 @@ namespace http::version::handler_1_0 {
         if (request.isContentTooLarge()) {
             setStatusMaybeErrorDoc(request, *pResponse, 413);
             return pResponse;
+        }
+
+        // Handle redirects
+        if (!conf::redirectRules.empty()) {
+            // Format raw path & break off query string
+            std::string rawPath = request.getPathStr();
+            const size_t queryIndex = rawPath.find('?');
+            const std::string queryString = queryIndex != std::string::npos ? rawPath.substr(queryIndex) : "";
+            while (queryIndex != std::string::npos && rawPath.size() > queryIndex)
+                rawPath.pop_back();
+
+            // Check redirect rule patterns
+            std::string locationBuf;
+            for (const std::unique_ptr<conf::Redirect>& pRedirect : conf::redirectRules) {
+                pRedirect->loadRedirectedPath(rawPath, locationBuf);
+                if (locationBuf.empty()) continue;
+
+                // New location found, set status (only 300-302 available for HTTP/1.0)
+                const unsigned int status = pRedirect->getStatus();
+                if (status > 302) ERROR_LOG << "HTTP/1.0 falling back from " << status << " status to 302 status" << std::endl;
+                pResponse->setStatus( status > 302 ? 302 : status );
+                pResponse->setHeader( "Location", locationBuf );
+                return pResponse;
+            }
         }
 
         // Verify path is restricted to document root