+++ src/http/version/handler_1_1.cpp
@@ -39,7 +39,7 @@ namespace http::version::handler_1_1 {
         if (!conf::matchConfigs.empty()) {
             // Format raw path & remove query string
             std::string rawPath = request.getPathStr();
-            const size_t queryIndex = rawPath.find('/');
+            const size_t queryIndex = rawPath.find('?');
             while (queryIndex != std::string::npos && rawPath.size() > queryIndex)
                 rawPath.pop_back();
 
@@ -68,6 +68,28 @@ namespace http::version::handler_1_1 {
         if (request.isContentTooLarge()) {
             setStatusMaybeErrorDoc(request, *pResponse, 413);
             return pResponse;
+        }
+
+        // Handle redirects
+        if (!conf::redirectRules.empty()) {
+            // Format raw path & break off query string
+            std::string rawPath = request.getPathStr();
+            const size_t queryIndex = rawPath.find('?');
+            const std::string queryString = queryIndex != std::string::npos ? rawPath.substr(queryIndex) : "";
+            while (queryIndex != std::string::npos && rawPath.size() > queryIndex)
+                rawPath.pop_back();
+
+            // Check redirect rule patterns
+            std::string locationBuf;
+            for (const std::unique_ptr<conf::Redirect>& pRedirect : conf::redirectRules) {
+                pRedirect->loadRedirectedPath(rawPath, locationBuf);
+                if (locationBuf.empty()) continue;
+
+                // New location found, set status
+                pResponse->setStatus( pRedirect->getStatus() );
+                pResponse->setHeader( "Location", locationBuf );
+                return pResponse;
+            }
         }
 
         File file(request.getPathStr());